Reserve old cryptosuite value range.
The Data Integrity cryptosuite identifier values were an experimental
use that was abandoned in favor of another technique. The value range is
being reserved in case anyone happened to have used those values. If no
feedback is received about this change, the range will be freed at some
point in the future.